requirement.txt
django==1.8.3
djangorestframework==3.1.3
django-debug-toolbar==1.3.2